Revert "test: add tests for prebundleSvelteLibraries"
This reverts commit 174e1dfadff9eaddd358eb23c6d91b07414ceb43.
