--- packages/e2e-tests/prebundle-svelte-deps/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "e2e-tests-prebundle-svelte-deps",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "e2e-test-dep-svelte-api-only": "file:../_test_dependencies/svelte-api-only",
-    "e2e-test-dep-svelte-hybrid": "file:../_test_dependencies/svelte-hybrid",
-    "e2e-test-dep-svelte-nested": "file:../_test_dependencies/svelte-nested",
-    "e2e-test-dep-svelte-simple": "file:../_test_dependencies/svelte-simple"
-  },
-  "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "workspace:*",
-    "sass": "^1.56.0",
-    "svelte": "^3.52.0",
-    "svelte-preprocess": "^4.10.7",
-    "vite": "^3.2.3"
-  }
-}
