--- packages/e2e-tests/prebundle-svelte-deps/svelte.config.js
@@ -1,7 +1,0 @@
-import preprocess from 'svelte-preprocess';
-export default {
-	preprocess: [preprocess()],
-	vitePlugin: {
-		prebundleSvelteLibraries: true
-	}
-};
